Changed the type of List from int to long

+++ twitter4j-core/src/main/java/twitter4j/api/ListsResources.java
@@ -61,7 +61,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/get/lists/statuses">GET lists/statuses | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    ResponseList<Status> getUserListStatuses(int listId, Paging paging)
+    ResponseList<Status> getUserListStatuses(long listId, Paging paging)
             throws TwitterException;
 
     /**
@@ -105,7 +105,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/members/destroy">POST lists/members/destroy | Twitter Developers</a>
      * @since Twitter4J 2.1.0
      */
-    UserList destroyUserListMember(int listId, long userId) throws TwitterException;
+    UserList destroyUserListMember(long listId, long userId) throws TwitterException;
 
     /**
      * @deprecated use {@link #destroyUserList(int)} instead
@@ -231,7 +231,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/get/lists/subscribers">GET lists/subscribers | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    PagableResponseList<User> getUserListSubscribers(int listId, long cursor) throws TwitterException;
+    PagableResponseList<User> getUserListSubscribers(long listId, long cursor) throws TwitterException;
 
     /**
      * Returns the subscribers of the specified list.
@@ -271,7 +271,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/subscribers/create">POST lists/subscribers/create | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    UserList createUserListSubscription(int listId) throws TwitterException;
+    UserList createUserListSubscription(long listId) throws TwitterException;
 
     /**
      * Make the authenticated user follow the specified list.
@@ -311,7 +311,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/get/lists/subscribers/show">GET lists/subscribers/show | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    User showUserListSubscription(int listId, long userId) throws TwitterException;
+    User showUserListSubscription(long listId, long userId) throws TwitterException;
 
     /**
      * Check if the specified user is a subscriber of the specified list.
@@ -353,7 +353,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/subscribers/destroy">POST lists/subscribers/destroy | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    UserList destroyUserListSubscription(int listId) throws TwitterException;
+    UserList destroyUserListSubscription(long listId) throws TwitterException;
 
     /**
      * Unsubscribes the authenticated user form the specified list.
@@ -391,7 +391,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/members/create_all">POST lists/members/create_all | Twitter Developers</a>
      * @since Twitter4J 3.0.0
      */
-    UserList createUserListMembers(int listId, long[] userIds) throws TwitterException;
+    UserList createUserListMembers(long listId, long[] userIds) throws TwitterException;
 
     /**
      * @deprecated use {@link #createUserListMembers(int, long[])} instead
@@ -439,7 +439,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/members/create_all">POST lists/members/create_all | Twitter Developers</a>
      * @since Twitter4J 2.1.7
      */
-    UserList createUserListMembers(int listId, String[] screenNames) throws TwitterException;
+    UserList createUserListMembers(long listId, String[] screenNames) throws TwitterException;
 
     /**
      * @deprecated use {@link #createUserListMembers(int, String[])} instead
@@ -487,7 +487,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/get/lists/members/show">GET lists/members/show | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    User showUserListMembership(int listId, long userId) throws TwitterException;
+    User showUserListMembership(long listId, long userId) throws TwitterException;
 
     /**
      * Check if a user is a member of the specified list.<br>
@@ -530,7 +530,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/get/lists/members">GET lists/members | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    PagableResponseList<User> getUserListMembers(int listId, long cursor)
+    PagableResponseList<User> getUserListMembers(long listId, long cursor)
             throws TwitterException;
 
     /**
@@ -574,7 +574,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/members/create">POST lists/members/create | Twitter Developers</a>
      * @since Twitter4J 2.1.0
      */
-    UserList createUserListMember(int listId, long userId) throws TwitterException;
+    UserList createUserListMember(long listId, long userId) throws TwitterException;
 
     /**
      * @deprecated use {@link #createUserListMember(int, long)}} instead
@@ -624,7 +624,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/destroy">POST lists/destroy | Twitter Developers</a>
      * @since Twitter4J 2.1.0
      */
-    UserList destroyUserList(int listId) throws TwitterException;
+    UserList destroyUserList(long listId) throws TwitterException;
 
     /**
      * Deletes the specified list. Must be owned by the authenticated user.
@@ -665,7 +665,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/post/lists/update ">POST lists/update | Twitter Developers</a>
      * @since Twitter4J 2.1.0
      */
-    UserList updateUserList(int listId, String newListName, boolean isPublicList, String newDescription)
+    UserList updateUserList(long listId, String newListName, boolean isPublicList, String newDescription)
             throws TwitterException;
 
     /**
@@ -727,7 +727,7 @@ as returned to in the response body's next_cursor and previous_cursor attributes
      * @see <a href="https://dev.twitter.com/docs/api/1.1/get/lists/show">https://dev.twitter.com/docs/api/1.1/get/lists/show | Twitter Developers</a>
      * @since Twitter4J 2.2.3
      */
-    UserList showUserList(int listId) throws TwitterException;
+    UserList showUserList(long listId) throws TwitterException;
 
     /**
      * Show the specified list. Private lists will only be shown if the authenticated user owns the specified list.
